Code cleanup + style cop. Checking in for demo
