+++ Samples/DeviceLab/App.xaml.cs
@@ -1,4 +1,9 @@
-﻿using System.Windows;
+﻿//----------------------------------------------------------------------------------------------
+// <copyright file="App.xaml.cs" company="Microsoft Corporation">
+//     Licensed under the MIT License. See LICENSE.TXT in the project root license information.
+// </copyright>
+//----------------------------------------------------------------------------------------------
+using System.Windows;
 
 namespace DeviceLab
 {
