+++ Samples/DeviceLab/Controls/NumberEntryBox.xaml
