+++ Samples/DeviceLab/ViewModels/DevicePortalCommandModel.cs
