+++ Samples/DeviceLab/ViewModels/DevicePortalViewModel.cs
