+++ Samples/DeviceLab/ViewModels/DeviceSignInViewModel.cs
