+++ Samples/DeviceLab/ViewModels/DiagnosticOutputViewModel.cs
