+++ Samples/DeviceLab/ViewModels/MainViewModel.cs
