+++ Samples/DeviceLab/Views/BindablePassword.xaml
