+++ Samples/DeviceLab/Views/BindablePassword.xaml.cs
