+++ Samples/DeviceLab/Views/DeviceCollectionView.xaml
