+++ Samples/DeviceLab/Views/DeviceCollectionView.xaml.cs
