+++ Samples/DeviceLab/Views/DeviceSignInView.xaml
