+++ Samples/DeviceLab/Views/DeviceSignInView.xaml.cs
